+++ webserver/app/models/registry.py
@@ -1,9 +1,13 @@
+import base64
+import json
 import re
+from kubernetes.client.exceptions import ApiException
 from sqlalchemy import Column, Integer, String, Boolean
 
-from app.helpers.const import DEFAULT_NAMESPACE
+from app.helpers.const import DEFAULT_NAMESPACE, TASK_NAMESPACE, TASK_PULL_SECRET_NAME
 from app.helpers.container_registries import AzureRegistry, BaseRegistry, DockerRegistry, GitHubRegistry
 from app.helpers.db import BaseModel, db
+from app.helpers.exceptions import ContainerRegistryException
 from app.helpers.kubernetes import KubernetesClient
 
 
@@ -56,7 +60,42 @@ class Registry(db.Model, BaseModel):
             },
             namespaces=[DEFAULT_NAMESPACE]
         )
+        self.update_regcred()
         super().add(commit)
+
+    def update_regcred(self):
+        """
+        Every time a new registry is added, we should propagate
+        the changes to the regcred secret in the task namespace
+        so we can pull those images
+        """
+        v1 = KubernetesClient()
+        try:
+            regcred = v1.read_namespaced_secret(TASK_PULL_SECRET_NAME, TASK_NAMESPACE)
+        except ApiException as kae:
+            if kae.status == 404:
+                regcred = v1.create_secret(
+                    name=TASK_PULL_SECRET_NAME,
+                    values={
+                        ".dockerconfigjson": json.dumps({"auths" : {}})
+                    },
+                    namespaces=[TASK_NAMESPACE],
+                    type='kubernetes.io/dockerconfigjson'
+                )
+            else:
+                raise ContainerRegistryException("Something went wrong while updating secrets")
+
+        dockerjson = json.loads(v1.decode_secret_value(regcred.data['.dockerconfigjson']))
+        dockerjson['auths'].update({
+            self.url: {
+                "username": self.username,
+                "password": self.password,
+                "email": "",
+                "auth": base64.b64encode(f"{self.username}:{self.password}".encode()).decode()
+            }
+        })
+        regcred.data['.dockerconfigjson'] = v1.encode_secret_value(json.dumps(dockerjson))
+        v1.patch_namespaced_secret(namespace=TASK_NAMESPACE, name=TASK_PULL_SECRET_NAME, body=regcred)
 
     def _get_creds(self):
         if hasattr(self, "username") and hasattr(self, "password"):